requirements.txt
streamlit
pandas
requests
beautifulsoup4
ddgs
openpyxl
psycopg2-binary
